Add files via upload
this is part of the backend for  a game called 8 minute empire. My group members, chen chen, Tino, and Tahar and I worked on this. My involvement consisted mainly in the cards, cards driver, and map files. Everyone else did their respective part and Tahar and Tino also helped with the map part as well.